backend/requirements.txt
asgiref==3.8.1
better-profanity==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
